+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Fixed an issue where **validate** did not fail on invalid playbook entities' versions (i.e. subplaybooks or scripts with higher fromversion than their parent playbook).
 * Added a new validation that checks whether a pack should be hidden.
 * Added a new format to `pack_metadata.json` to hide a pack that should be hidden.
 * Added support for running lint via a remote docker ssh connection. Use `DOCKER_HOST` env variable to specify a remote docker connection, such as: `DOCKER_HOST=ssh://myuser@myhost.com`.
@@ -22,6 +23,7 @@
 * Added a check to the **lint** command for finding hard-coded usage of the http protocol.
 * Locked the dependency on Docker.
 * Removed a traceback line from the **init** command templates: BaseIntegration, BaseScript.
+* Updated the token in **_add_pr_comment** method from the content-bot token to the xsoar-bot token.
 
 ## 1.6.7